Respect `callback_url` param to _logout_route
Before this change, the query parameter `callback_url`
in MSALAuthorization._logout_route was being ignored
by the logic. I verified with the author that this was
not intentional.

This change now prefers the query parameter first,
ahead of the Referer header and the configured default.

Another alternative would have been to remove the
query parameter, as it was not currently being
acknowledged anyway.

- https://github.com/dudil/fastapi_msal/issues/45